requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.31
PyYAML==6.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
